add RenderBalanceText and RenderStatsText component
